+++ shared-components/backend/lib/config/config.go
@@ -1,0 +1,9 @@
+package config
+
+import (
+	"tmossDev.github.com/eco-system/shared-components/backend/lib/config/types"
+)
+
+type Config interface {
+	GetConfig(configName string) (*types.ConfigModel, error)
+}