+++ shared-components/backend/lib/constants/general.go
@@ -1,0 +1,6 @@
+package constants
+
+const (
+	CTXRequestIdKey  = "request-id"
+	DefaultRequestId = "default-request-id"
+)