+++ shared-components/backend/lib/datastore/datastore.go
@@ -1,0 +1,7 @@
+package datastore
+
+type DataStore interface {
+	Connect() error
+	Close() error
+	Ping() error
+}