+++ shared-components/backend/lib/env/constants/default.go
@@ -1,0 +1,6 @@
+package constants
+
+const (
+	DefaultPort    = "8080"
+	DefaultVersion = "v1.0.0"
+)