+++ shared-components/backend/lib/env/constants/env.go
@@ -1,0 +1,7 @@
+package constants
+
+const (
+	Port       = "PORT"
+	SecretName = "SECRET_NAME"
+	Env        = "ENV"
+)