+++ shared-components/backend/lib/transport/http/constants/default.go
@@ -1,0 +1,9 @@
+package constants
+
+const (
+	URLPageSizeKey = "pageSize"
+	URLPageKey     = "page"
+
+	DefaultPageSize = 10
+	DefaultIndex    = 0
+)