+++ shared-components/backend/lib/transport/http/constants/http.go
@@ -1,0 +1,12 @@
+package constants
+
+const (
+	ApiPrefix = "/api"
+)
+
+const (
+	ContentType      = "Content-Type"
+	ApplicationJson  = "application/json"
+	ErrFormatLogging = "returned error: %s"
+	RspFormatLogging = "response out: %s"
+)