+++ shared-components/backend/lib/transport/http/constants/middleware.go
@@ -1,0 +1,5 @@
+package constants
+
+const (
+	CTXRequestIdKey = "request-id"
+)