+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Project Change Log
 
+## v1.1.0 - (4 Changes)
+- Added shared-components for common components.
+- Added GoLang common libraries to shared-components/backend/lib.
+- Added user-service shell application for user management internal calls.
+- Added user-gateway shell application for user management external calls.
+
 ## v1.0.0 - (4 Changes)
 - Added foundation use case for getting local postgres database up and running.
 - Added database Init Script for starting up the shop database from scratch